requirements.txt
fastapi
uvicorn
python-multipart
requests
beautifulsoup4
pypdf2
langchain
langchain-community
faiss-cpu
sentence-transformers
datasets
huggingface_hub
python-dotenv
flask